movebackforth: use Laura's marker follower
We reuse a component by group 3 for following the marker path. Adjust
"The Boss" to use this service instead of waiting for the marker_go.

This might be a nice example of code reuse in the Research Camp. This
external component was also much easier to integrate than re-using the
coordination and computation sub-components that were provided in the
MarkerBasedNavigation stack.